Add native like up and down movement with arrow keys
