+++ web/src/lib/caretHelpers.ts
@@ -19,10 +19,10 @@ const isCaretAtTop = (element: HTMLElement) => {
   const elementPosition = element.getBoundingClientRect().top + topPadding;
   const caretPosition = y - elementPosition;
 
-  console.log(caretPosition);
+  console.log(caretPosition, caretPosition < 5);
 
   // ~ Check if the caret is at the top of the element (within 5px)
-  return caretPosition === 0 || caretPosition < 5;
+  return caretPosition < 5;
 }
 
 const isCaretAtBottom = (element: HTMLElement) => {
@@ -41,10 +41,10 @@ const isCaretAtBottom = (element: HTMLElement) => {
 
   const caretPosition = (elementPosition - y) - getStyle(element, 'line-height');
 
-  console.log(caretPosition);
+  console.log(caretPosition, caretPosition < 0);
 
   // ~ Check if the caret is at the bottom of the element (within 5px)
-  return caretPosition === 0 || caretPosition > -5;
+  return caretPosition < 5;
 }
 
 export { isCaretAtTop, isCaretAtBottom };
