+++ web/src/lib/pages/updatePage.ts
@@ -2,6 +2,7 @@
 /* eslint-disable no-underscore-dangle */
 import crypto from 'crypto';
 
+import focusBlockAtIndex from '../focusBlockAtIndex';
 import type PageDataInterface from '../../types/pageTypes';
 import SaveManager from '../../classes/SaveManager';
 
@@ -90,39 +91,7 @@ const removeBlock = async (
   // ~ Wait for the page to update to focus the end of the previous block
   await new Promise((resolve) => setTimeout(resolve, 5));
 
-  /**
-   * Select the end of the element
-   * @param element The element to focus
-   */
-  const selectEnd = (element: HTMLElement) => {
-    element.focus();
-    // ~ Move the cursor to the end of the block unless the only text is a newline
-    if (element.textContent === '\n') return;
-
-    const range = document.createRange();
-    const sel = window.getSelection();
-    range.setStart(element, element.childNodes.length);
-    range.collapse(true);
-    sel?.removeAllRanges();
-    sel?.addRange(range);
-  };
-
-  // ~ Find the previous editable block
-  while (index > 0) {
-    index -= 1;
-
-    const block = document.getElementById(tempPageData.message.data[index]._id);
-    if (block?.getAttribute('contenteditable') === 'true') {
-      selectEnd(block);
-      return;
-    }
-  }
-
-  // ~ If there is no previous editable block, focus the first editable block
-  const block = document.getElementById('page-title')?.firstChild;
-  if (!block) return;
-
-  selectEnd(block as HTMLElement);
+  focusBlockAtIndex(index, tempPageData);
 };
 
 const editBlock = async (
